requirements.txt
numpy
matplotlib
numba
pygame-ce>=2.5.4
pygame_gui
jplephem
hypothesis
